+++ src/main/java/com/volmit/react/api/DTest.java
@@ -59,6 +59,6 @@ public class DTest
 			score += b + 1;
 		}
 
-		return 1100 + (int) (score);
+		return 1100 + score;
 	}
 }
