+++ src/main/java/com/volmit/react/util/VectorMath.java
@@ -485,7 +485,7 @@ public class VectorMath
 		CDou cy = new CDou(360);
 		cy.set(y);
 		l.setYaw((float) cy.get());
-		l.setPitch((float) Math.max(-90, p - amt));
+		l.setPitch(Math.max(-90, p - amt));
 
 		return l.getDirection();
 	}
@@ -508,7 +508,7 @@ public class VectorMath
 		CDou cy = new CDou(360);
 		cy.set(y);
 		l.setYaw((float) cy.get());
-		l.setPitch((float) Math.min(90, p + amt));
+		l.setPitch(Math.min(90, p + amt));
 
 		return l.getDirection();
 	}
